pyproject.toml
[project]
name = "pyclaw"
version = "0.1.0"
description = "PyClaw - Python personal AI assistant built on LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.2",
    "langgraph-checkpoint-sqlite>=2.0",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=4.0",
    "langchain-ollama>=0.3",
    "typer[all]>=0.15",
    "rich>=13.0",
    "pydantic>=2.0",
    "apscheduler>=3.10,<4",
    "tavily-python>=0.7.21",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3"]
slack = ["slack-bolt>=1.18"]
search = ["duckduckgo-search>=6.0"]
all = [
    "python-telegram-bot>=21.0",
    "discord.py>=2.3",
    "slack-bolt>=1.18",
    "duckduckgo-search>=6.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
pyclaw = "pyclaw.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyclaw"]
